+++ src/main/java/org/scion/jpan/ScionService.java
@@ -128,9 +128,9 @@ public class ScionService {
       segmentStub = SegmentLookupServiceGrpc.newBlockingStub(channel);
     }
     shutdownHook = addShutdownHook();
-    Shim.install(this);
     try {
       getLocalIsdAs(); // Init
+      checkStartShim();
     } catch (RuntimeException e) {
       // If this fails for whatever reason we want to make sure that the channel is closed.
       try {
@@ -139,6 +139,16 @@ public class ScionService {
         // Ignore, we just want to get out.
       }
       throw e;
+    }
+  }
+
+  private void checkStartShim() {
+    // Start SHIM unless we have port range 'ALL'. However, config overrides this setting.
+    String config = ScionUtil.getPropertyOrEnv(Constants.PROPERTY_SHIM, Constants.ENV_SHIM);
+    boolean hasAllPorts = getLocalPortRange().hasPortRangeALL();
+    boolean start = config != null ? Boolean.parseBoolean(config) : !hasAllPorts;
+    if (start) {
+      Shim.install(this);
     }
   }